+++ src/Products/MeetingCommunes/profiles/simple/import_data.py
@@ -98,7 +98,6 @@ simpleMeeting.powerObservers = (
 simpleMeeting.workflowAdaptations = [
     'no_publication', 'refused', 'accepted_but_modified', 'delayed',
     'return_to_proposing_group', 'only_creator_may_delete', 'pre_accepted']
-simpleMeeting.transitionsForPresentingAnItem = ('propose', 'validate', 'present', )
 simpleMeeting.onMeetingTransitionItemActionToExecute = (
     {'meeting_transition': 'freeze',
      'item_action': 'itemfreeze',