+++ CHANGES.rst
@@ -10,7 +10,11 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
 
 - Rename "Commission des volontaires" profile to "Bureau des volontaires".
   [aduchene]
-- Add helper print method to be able to group by custom method instead of persistent value on item. This method must begin by "_group_by_"
+- Add helper print method to be able to group by custom method instead of persistent value on item.
+  This method must begin by "_group_by_".
+  [anuyens, gbastien]
+- Field `MeetingConfig.transitionsForPresentingAnItem` was removed, adapted profiles accordingly.
+  [gbastien]
 
 4.2b21 (2022-06-14)
 -------------------